# Update SqliteCommand.Prepare implementation