--- a/src/Microsoft.Data.Sqlite.Core/Properties/Resources.resx
+++ b/src/Microsoft.Data.Sqlite.Core/Properties/Resources.resx
@@ -198,4 +198,10 @@
   <data name="CannotStoreNaN" xml:space="preserve">
     <value>Cannot store 'NaN' values.</value>
   </data>
+  <data name="DataReaderOpen" xml:space="preserve">
+    <value>An open reader is already associated with this command. Close it before opening a new one.</value>
+  </data>
+  <data name="SetRequiresNoOpenReader" xml:space="preserve">
+    <value>An open reader is associated with this command. Close it before changing the {propertyName} property.</value>
+  </data>
 </root>