… .gitignore
… src/Microsoft.Data.Sqlite.Core/SqliteDataReader.cs
… test/Microsoft.Data.Sqlite.Tests/SqliteCommandTest.cs