https://github.com/aspnet/Microsoft.Data.Sqlite/pull/388/commits/8b37cabb5214eab12cba49fd915d965c3bad9ca0 improves on the previous implementation we tried.

I ran a quick benchmark reusing a parameterized `INSERT` command for fun: 4x faster!

&nbsp; | Mean | Error | StdDev
--- | ---:| ---:| ---:
**Before** | 4.184 us | 0.0835 us | 0.2001 us
**After** | 1.048 us | 0.0208 us | 0.0415 us

Resolves #235 

/cc @AlexanderTaeschner